refactor: split music.rs into module with ncm_api backend
- Split src/skills/music.rs into src/skills/music/ module
- ts3audiobot.rs: existing TS3AudioBot private-message backend
- tsbot_http.rs: existing NeteaseTSBot HTTP backend
- ncm_api.rs: new NeteaseCloudMusicApi backend (search + play)
- response.rs: MusicBackendResponse with __play_url support
- command.rs: reserved for future MusicCommand enum
- mod.rs: unified dispatch with ncm_api backend support
