+++ src/skills/music/command.rs
@@ -1,0 +1,1 @@
+// Unified music command types (reserved for future use).
